requirements.txt
python-chess
Pillow
pygame
# tkinter is included with Python standard library 